requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════
# SPJ POS v13.30 — requirements.txt (VERSIÓN CORREGIDA)
# ═══════════════════════════════════════════════════════════════════════════════

# ───────────────────────────────────────────────────────────────────────────────
# CORE — PyQt5 y GUI
# ───────────────────────────────────────────────────────────────────────────────
PyQt5==5.15.11
PyQt5-sip==12.18.0
PyQtWebEngine==5.15.7          # ← VERSIÓN MÁXIMA DISPONIBLE (no existe 5.15.9)

# ───────────────────────────────────────────────────────────────────────────────
# BASE DE DATOS
# ───────────────────────────────────────────────────────────────────────────────
# sqlite3 viene con Python — no requiere instalación

# ───────────────────────────────────────────────────────────────────────────────
# WHATSAPP MICROSERVICE — FastAPI
# ───────────────────────────────────────────────────────────────────────────────

fastapi==0.115.*
uvicorn[standard]==0.34.*
httpx==0.28.*
python-dotenv==1.0.*
apscheduler==3.10.*

# ───────────────────────────────────────────────────────────────────────────────
# HARDWARE — Impresoras, Básculas, Scanners
# ───────────────────────────────────────────────────────────────────────────────
pyserial>=3.5
python-escpos>=3.0a10
pywin32>=306; sys_platform == 'win32'

# ───────────────────────────────────────────────────────────────────────────────
# IMÁGENES Y CÓDIGOS QR
# ───────────────────────────────────────────────────────────────────────────────
Pillow>=10.0.0
qrcode[pil]>=7.4.2

# ───────────────────────────────────────────────────────────────────────────────
# PROCESAMIENTO DE DATOS — Forecast y BI
# ───────────────────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# ───────────────────────────────────────────────────────────────────────────────
# REPORTES Y EXPORTACIÓN
# ───────────────────────────────────────────────────────────────────────────────
reportlab>=4.0.0
openpyxl>=3.1.0
FPDF2>=2.7.0

# ───────────────────────────────────────────────────────────────────────────────
# PAGOS — MercadoPago
# ───────────────────────────────────────────────────────────────────────────────
mercadopago>=2.0.0

# ───────────────────────────────────────────────────────────────────────────────
# SEGURIDAD Y AUTENTICACIÓN
# ───────────────────────────────────────────────────────────────────────────────
bcrypt>=4.0.0
cryptography>=41.0.0

# ───────────────────────────────────────────────────────────────────────────────
# MATCHING DE PRODUCTOS — Fuzzy Search
# ───────────────────────────────────────────────────────────────────────────────
python-Levenshtein>=0.21.0

# ───────────────────────────────────────────────────────────────────────────────
# UTILIDADES
# ───────────────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0

# ═══════════════════════════════════════════════════════════════════════════════
# INSTALACIÓN
# ═══════════════════════════════════════════════════════════════════════════════
# pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════